Revert "refactor: test 패키지 구조 수정"
This reverts commit 8fd90d362e53e6a459a34db0cbab5ca566ee2a24.
