+++ src/main/java/com/gistpetition/api/petition/domain/Agreements.java
@@ -17,7 +17,7 @@ public class Agreements {
     @OneToMany(mappedBy = "petition", cascade = CascadeType.PERSIST, orphanRemoval = true)
     private final List<Agreement> agreements = new ArrayList<>();
 
-    public Agreements() {
+    protected Agreements() {
     }
 
     public void add(Agreement agreement) {
