+++ src/test/java/com/gistpetition/api/IntegrationTest.java
@@ -1,4 +1,4 @@
-package com.gistpetition.api.integration;
+package com.gistpetition.api;
 
 import com.gistpetition.api.utils.emailsender.EmailSender;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -12,5 +12,4 @@ import org.springframework.test.context.jdbc.Sql;
 public abstract class IntegrationTest {
     @MockBean
     EmailSender emailSender;
-
 }
