+++ src/test/java/com/gistpetition/api/petition/domain/CategoryTest.java
@@ -1,7 +1,6 @@
-package com.gistpetition.api.unit.petition;
+package com.gistpetition.api.petition.domain;
 
 import com.gistpetition.api.exception.petition.NoSuchCategoryException;
-import com.gistpetition.api.petition.domain.Category;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
