+++ src/test/java/com/gistpetition/api/user/application/LoginServiceTest.java
@@ -1,9 +1,8 @@
-package com.gistpetition.api.integration.user;
+package com.gistpetition.api.user.application;
 
+import com.gistpetition.api.IntegrationTest;
 import com.gistpetition.api.exception.user.NoSuchUserException;
 import com.gistpetition.api.exception.user.NotMatchedPasswordException;
-import com.gistpetition.api.integration.IntegrationTest;
-import com.gistpetition.api.user.application.LoginService;
 import com.gistpetition.api.user.domain.SimpleUser;
 import com.gistpetition.api.user.domain.User;
 import com.gistpetition.api.user.domain.UserRepository;
