+++ src/test/java/com/gistpetition/api/user/application/UserServiceTest.java
@@ -1,11 +1,10 @@
-package com.gistpetition.api.integration.user;
+package com.gistpetition.api.user.application;
 
+import com.gistpetition.api.IntegrationTest;
 import com.gistpetition.api.exception.user.DuplicatedUserException;
 import com.gistpetition.api.exception.user.InvalidEmailFormException;
 import com.gistpetition.api.exception.user.NoSuchUserException;
 import com.gistpetition.api.exception.user.NotMatchedPasswordException;
-import com.gistpetition.api.integration.IntegrationTest;
-import com.gistpetition.api.user.application.UserService;
 import com.gistpetition.api.user.domain.User;
 import com.gistpetition.api.user.domain.UserRepository;
 import com.gistpetition.api.user.domain.UserRole;
